+++ srcs/functions/func_lldi.c
@@ -6,7 +6,7 @@
 /*   By: acoupleu <acoupleu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/05/04 15:42:36 by acoupleu          #+#    #+#             */
-/*   Updated: 2017/05/10 18:32:44 by acoupleu         ###   ########.fr       */
+/*   Updated: 2017/05/15 21:13:22 by acoupleu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,7 +50,7 @@ void	lindirect_load(t_map *map, t_process *proc)
 				return ;
 			}
 			result = (pos + ((proc->reg[(int)map->arena[(pos + 2) % MEM_SIZE] - 1]) % 3)) % MEM_SIZE;
-			result += hex_to_int(0x00, 0x00, map->arena[(pos + 3) % MEM_SIZE], map->arena[(pos + 4) % MEM_SIZE]);
+			result += (short)hex_to_int(0x00, 0x00, map->arena[(pos + 3) % MEM_SIZE], map->arena[(pos + 4) % MEM_SIZE]);
 			proc->reg[(int)map->arena[(pos + 5) % MEM_SIZE] - 1] = read_in_arena(map, (pos + result) % MEM_SIZE);
 			proc->pc = proc->pc + 6;
 		}
@@ -62,7 +62,7 @@ void	lindirect_load(t_map *map, t_process *proc)
 				fail_func(proc, 6, 1);
 				return ;
 			}
-			result = (pos + ((hex_to_int(0x00, 0x00, map->arena[(pos + 2) % MEM_SIZE], map->arena[(pos + 3) % MEM_SIZE])) % 3)) % MEM_SIZE;
+			result = (pos + ((short)hex_to_int(0x00, 0x00, map->arena[(pos + 2) % MEM_SIZE], map->arena[(pos + 3) % MEM_SIZE]) % 3)) % MEM_SIZE;
 			result += proc->reg[(int)map->arena[(pos + 4) % MEM_SIZE] - 1];
 			proc->reg[(int)map->arena[(pos + 5) % MEM_SIZE] - 1] = read_in_arena(map, (pos + result) % MEM_SIZE);
 			proc->pc = proc->pc + 6;
@@ -74,8 +74,8 @@ void	lindirect_load(t_map *map, t_process *proc)
 				fail_func(proc, 7, 1);
 				return ;
 			}
-			result = (pos + ((hex_to_int(0x00, 0x00, map->arena[(pos + 2) % MEM_SIZE], map->arena[(pos + 3) % MEM_SIZE])) % 3 )) % MEM_SIZE;
-			result += hex_to_int(0x00, 0x00, map->arena[(pos + 4) % MEM_SIZE], map->arena[(pos + 5) % MEM_SIZE]);
+			result = (pos + ((short)hex_to_int(0x00, 0x00, map->arena[(pos + 2) % MEM_SIZE], map->arena[(pos + 3) % MEM_SIZE]) % 3 )) % MEM_SIZE;
+			result += (short)hex_to_int(0x00, 0x00, map->arena[(pos + 4) % MEM_SIZE], map->arena[(pos + 5) % MEM_SIZE]);
 			proc->reg[(int)map->arena[(pos + 6) % MEM_SIZE] - 1] = read_in_arena(map, (pos + result) % MEM_SIZE);
 			proc->pc = proc->pc + 7;
 		}
